backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
sqlmodel>=0.0.14
psycopg2-binary>=2.9.9
pydantic>=2.5.0
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
celery[redis]>=5.3.0
mlflow>=2.8.0
torch>=2.0.0
pandas>=2.1.3
numpy>=1.26.2
scipy>=1.11.4
scikit-learn>=1.3.2
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.1
httpx>=0.25.0
sdv>=1.15.0
opacus>=1.4.0
python-multipart>=0.0.6

# LLM Integration
google-generativeai>=0.3.0
groq>=0.4.0
redis>=5.0.0
tiktoken>=0.5.0

# Database Migrations
alembic>=1.13.0

# Report Export
weasyprint>=60.0
python-docx>=1.0.0
markdown2>=2.4.0

# AWS S3 Storage
boto3>=1.34.0
botocore>=1.34.0
